requirements.txt
awsebcli==3.17.0
backcall==0.1.0
blessed==1.17.0
boto3==1.11.7
botocore==1.13.50
cached-property==1.5.1
cement==2.8.2
certifi==2019.11.28
chardet==3.0.4
Click==7.0
colorama==0.3.9
cycler==0.10.0
decorator==4.4.1
docker==3.7.3
docker-compose==1.23.2
docker-pycreds==0.4.0
dockerpty==0.4.1
docopt==0.6.2
docutils==0.15.2
Flask==1.1.1
Flask-Cors==3.0.8
Flask-SQLAlchemy==2.4.1
future==0.16.0
gunicorn==20.0.4
idna==2.7
imageio==2.6.1
ipython==7.11.1
ipython-genutils==0.2.0
itsdangerous==1.1.0
jedi==0.15.2
Jinja2==2.10.3
jmespath==0.9.4
joblib==0.14.1
jsonschema==2.6.0
kiwisolver==1.1.0
MarkupSafe==1.1.1
matplotlib==3.1.2
networkx==2.4
numpy==1.18.1
opencv-python==4.1.2.30
pandas==0.25.3
parso==0.5.2
pathspec==0.5.9
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.0.0
prompt-toolkit==3.0.2
psycopg2-binary==2.8.4
ptyprocess==0.6.0
Pygments==2.5.2
pyparsing==2.4.6
python-dateutil==2.8.0
python-decouple==3.3
pytz==2019.3
PyWavelets==1.1.1
PyYAML==3.13
requests==2.20.1
s3transfer==0.3.1
scikit-image==0.16.2
scikit-learn==0.21.3
scipy==1.4.1
semantic-version==2.5.0
six==1.11.0
SQLAlchemy==1.3.13
termcolor==1.1.0
texttable==0.9.1
traitlets==4.3.3
urllib3==1.24.3
wcwidth==0.1.8
websocket-client==0.57.0
Werkzeug==0.16.0
